requirements.txt
flask
python-telegram-bot==20.3  # Ensure the version matches your implementation